+++ core/render_scheduler_stub.go
@@ -1,0 +1,7 @@
+//go:build !js || !wasm
+
+package core
+
+func (c *HTMLComponent) requestRender() {}
+
+func cancelComponentRender(string) {}